move tasks in part2
