+++ src/main/java/part2/Task4.java
@@ -1,3 +1,5 @@
+package part2;
+
 import java.io.BufferedReader;
 import java.io.IOException;
 import java.io.InputStream;
